Chi Square
This was done as a proof of concept for my mathematics Internal Assessment for IB in grade 11. The code  uses the concept of chi square in with respect to the average occurrence of each letter in the english alphabet to crack some simple ciphers. 

CaesarShift class is the test case for the program used to crack the a simple caesar shift cipher through frequency analysis. The output of the program will list the possible keys with the correct one with the lowest chi square number first. 

VigenereCipher class is used to crack a simple 4 length key vigenere cipher, the program does this by looking through all the possible keys, sorting the results by the starting with the lowest chi square number, then going through a dictionary to check that the output can potentially be an english word. Due to the larger number of outcomes from the vigenere cipher the correct key may not be the first output but should be within the first 50. 

CodeCalc - contains conversions from char-int, as well as the code for the dictionary check which checks to see if the string can fit a sequence of english words, using the following algorithm(simplified). 
1. take first 2 letters, use to get all words starting with those 2 letters. 
2. For each word check if string starts with word, if true proceed to next step 
3. Remove the portion of the string containing the word and repeat from step 2 until either a word matches the remaining string or all possible words in the list are exhausted

The dictionary check runs with moderate efficiency and only takes about a second to check against a dictionary of 479k english words. Due to the size of the dictionary used a false positive is likely.

ChiSquare is the class that handles the chi square calculation

EnigmaCipher is incomplete due to running out of time for the assignment
-I have created an unlisted side project which can cipher files based on the principles of the enigma machine, rotor(s), swapboard, and steps which advance the rotor by 1 or more step.